requirement.txt
google-cloud-spanner
google-auth